requirements.txt
torch
sentence-transformers
faiss-cpu
transformers
numpy